+++ api-service/src/api/spaces/list.rs
@@ -18,7 +18,7 @@ pub struct PaginationParams {
     #[validate(range(min = 1))]
     offset_id: Option<i32>,
     #[validate(range(min = 1))]
-    limit: Option<i64>,
+    limit: Option<i32>,
     user: Option<bool>,
     #[validate(length(min = 16, max = 32))]
     building_id: Option<String>,
@@ -28,7 +28,7 @@ pub struct PaginationParams {
 pub struct ListSpacesResponse {
     pub spaces: Vec<PublicSpace>,
     pub next_offset_id: Option<i32>,
-    pub limit: i64,
+    pub limit: i32,
 }
 
 // ======================================================================
@@ -64,7 +64,7 @@ pub async fn list_spaces(
         query.building_id.to_owned(),
     )
     .await?;
-    let next_offset_id = if spaces.len() as i64 == limit {
+    let next_offset_id = if spaces.len() as i32 == limit {
         spaces.last().map(|b| b.id)
     } else {
         None
